Tidy up anti-replay and late report checks
Among many other things about reports, Aggregators are required to
check that reports only get aggregated once and that no report be added
to a collected  batch. These two checks differ from the numerous other
per-report validations in that they can require expensive storage
queries and/or synchronization across many jobs. And while it's not OK
to include a report that fails either check into a batch bucket or
batch, there's no harm in preparing/refining/validating a report that
will eventually be rejected for either reason. Therefore, we don't want
to require that these checks be done as early as possible during
aggregation job initialization, but want to allow them to be run at any
time before they get added to a batch bucket.

The existing text for this was spread out across
{{input-share-validation}}, Leader agg job init, Leader agg job
continue, Helper agg job init and Helper agg job continue. In
particular, the Helper's anti-replay requirement appeared as a MAY in
one place but a MUST in another, which is confusing for implementors.

To address this, I beef up the notion of "committing an output share"
to mean the irreversible action of including an output share into a
batch bucket. This allows us to more crisply discuss it in the different
init and continue sections. We also move the anti-replay and late report
checks into {{batch-buckets}} (which was already pointed to by all the
relevant sections).

Reviewers might ask why #678 inlines some common validation logic into
multiple sections while this PR does the opposite. Subjectively, I feel
that {#batch-buckets} applies more neatly to the four places that we
refer to it from than did the {#batch-validation} section removed in
PR #678.

Closes #670