Use HTTP message content length where possible
In #686, we use the length of the enclosing HTTP message to figure out
how long protocol messages are. For consistency, we now do the same
thing for:

- AggregationJobInitReq
- AggregationJobResp
- AggregationJobContinueReq
- CollectionJobReq
- AggregateShareReq

In the last case, the message fields had to be re-ordered so that
`agg_param` comes last.

Additionally, the word "denotes" was misused in several places. For
example, we had:

```
struct {
  Report reports[message_length];
} UploadRequest;

Here `message_length` denotes the length of the HTTP message content...
```

Per Merriam-Webster, to denote means "to serve as an indication of"
([1]). Implementations are not meant to use the length of `reports` as
an indication of how long the HTTP message content is. It's the other
way around. Let's just say "is" and avoid this trouble.

[1]: https://www.merriam-webster.com/dictionary/denote

Closes #714