Don't check timestamp in input share validation
Report timestamps are in `ReportMetadata`, which is in the unencrypted
portion of the report. Therefore the well-formedness of timestamps is
already covered by the check that the enclosing `AggregationJobInitReq`
is valid, and need not be repeated during input share validation. In
particular, we must make it clear that an invalid timestamp should cause
the entire `AggregationJobInitReq` to be rejected, not the individual
report.